+++ Tests/Test-FailSkip.ps1
@@ -1,0 +1,22 @@
+function Test-FailSkip {
+    [CmdletBinding()]
+    param (
+        [Parameter(Mandatory, ValueFromPipeline)]
+        [Alias("InputObject")]
+        $SomeParameter,
+        [Parameter(ValueFromRemainingArguments)]
+        $SomethingLeftover
+    )
+
+    begin {
+    }
+    process {
+        Start-Jojoba {
+            Write-JojobaFail "Broken"
+            Write-JojobaSkip "Skip because of $SomeParameter"
+        }
+    }
+    end {
+        Publish-Jojoba
+    }
+}